moving reranker to the handlers directory
