+++ tests/integrations/utilities/rag/rerankers/test_openai_reranker.py
@@ -1,5 +1,5 @@
 import pytest
-from mindsdb.integrations.utilities.rag.rerankers.openai_reranker import Reranker
+from mindsdb.integrations.utilities.handlers.api_utilities.openai.openai_reranker import Reranker
 
 DOCUMENT_EXAMPLE = """The former home secretary Suella Braverman acted unlawfully in making it easier for the police to criminalise peaceful protests, the high court has ruled. She was found to have both acted outside her powers and to have failed to consult properly over regulations that would be likely to increase prosecutions of protesters by a third. Hundreds of protesters have been arrested since the government redefined the sort of protest that could be restricted by the police, allowing it where there is merely a “more than minor” hindrance to people’s daily lives."""
 QUERY_EXAMPLE = "What happened with Suella Braverman?"
